% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phytree.R
\name{sim_phytree}
\alias{sim_phytree}
\title{Simulate a Phylogenetic Tree from a Time Tree}
\usage{
sim_phytree(timetree, murate = 0, theta = 0)
}
\arguments{
\item{timetree}{An object of class \code{phylo} (from the \code{ape} package).}

\item{murate}{Numeric. Mutation rate used to scale the time tree branch lengths.}

\item{theta}{Numeric. Population size parameter for the coalescent simulation.
If 0, a simple scaling is performed.}
}
\value{
An object of class \code{phylo} representing the simulated gene tree.
}
\description{
Converts an epidemiological time tree into a phylogenetic tree by scaling
branch lengths by a mutation rate and optionally simulating the coalescent
process using the multispecies coalescent framework.
}
