% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_betnet.R
\name{summary_betnet}
\alias{summary_betnet}
\title{Summary and Parameter Estimation for BetNet Output}
\usage{
summary_betnet(
  betnet_output_file,
  burnin = 0.1,
  onsite_time,
  removal_time,
  plot = TRUE
)
}
\arguments{
\item{betnet_output_file}{String. Path to the CSV output from BetNet.}

\item{burnin}{Numeric. Proportion of initial iterations to discard (default 0.1).}

\item{onsite_time}{Numeric vector. Onsite times for cases in the outbreak.}

\item{removal_time}{Numeric vector. Removal times for cases in the outbreak.}

\item{plot}{Logical. If TRUE, displays MCMC trace plots.}
}
\value{
A list containing:
\itemize{
\item \code{theta}: Posterior mean of the theta parameter.
\item \code{mu}: Posterior mean of the mutation rate.
\item \code{infrate}: Posterior mean of the infection rate.
\item \code{transmission}: A data frame of the reconstructed transmission network.
}
}
\description{
Processes MCMC output from BetNet, generates convergence plots,
calculates posterior means for parameters, and reconstructs the
most likely transmission network.
}
